requirements.txt
fastapi[standard]
python-dotenv

pydantic[email,timezone]
sqlalchemy
alembic
oracledb
passlib[bcrypt]